kappa::dhs
Function toon_ackerman_1981

pub fn toon_ackerman_1981(dhsc: &DHSConfig<'_>) -> Result<DHSResult>

This function computes electromagnetic scattering by a stratified sphere, i.e., a particle consisting of a spherical core surrounded by a spherical shell. The surrounding medium is assumed to have refractive index unity. The formulas, manipulated to avoid the ill-conditioning that plagued earlier formulations, were published in: Toon, O. and T. Ackerman, Applied Optics 20, 3657 (1981).

Note

The refractive index entering into this routine use a convention where the imaginary part has a different singn than what is used in modern books.