kappa

Module fractal

Computes light scattering properties of randomly oriented fractal dust aggregates by means of the modified mean field theory developed in Tazaki & Tanaka (2018). This code is also capable of computing the light scattering solution based on the Rayleigh-Gans-Debye theory and the Mean field theory.

References

Based on OPFRACTAL v3.0 from optool by Ryo Tazaki.

Structs

FractalConfig

FractalResult

Enums

FractalCutoff

FractalGeometry

FractalSolver

Functions

get_geometric_cross_section_okuzumi 🔒

lorenz_mie 🔒

Calculate Lorenz-Mie scattering coefficients (an,bn) for a monomer particle.

maxwell_garnett_mixing 🔒

This function finds effective refractive index of an aggregates by using Maxwell-Garnett mixing rule.

mean_field 🔒

Solves the mean field theory.

mean_scattering

Computes light scattering properties of randomly oriented fractal dust aggregates by means of the modified mean field theory developed in Tazaki & Tanaka (2018). This code is also capable of computing the light scattering solution based on the Rayleigh-Gans-Debye theory and the Mean field theory.

renormalize 🔒

Calculate scattering amplitude S1, S2 from given scattering coefficients d.

structure_factor_fn 🔒

Integrand of the static structure factor.

structure_factor_integration 🔒

This subroutine performs integration of the static structure factor by Botet et al. (1995) based on df-dependent cut-off function for two-point correlation function. The statis structure factor S(q) is given by