kappa::fractal

Function mean_field

fn mean_field(
    fracc: &FractalConfig,
    ad: &CMatrix,
    x_g: f64,
    nmax: usize,
) -> Result<CVector>

Solves the mean field theory.

PART I.

Calculate a(nu,n,p) and b(nu,n,p):

               2p + 1  /+1
   a(nu,n,p) = -------- | dx P_nu^1(x) * P_n^1(x) * P_p(x),
                  2     /-1
 
                2p + 1  /+1                         dP_p(x)
   b(nu,n,p) = -------- | dx P_nu^1(x) * P_n^1(x) * -------,
                  2     /-1                            dx

where P_n^m is the associated Legendre function (n: degree, m: order), P_n is the Legendre polynominal function (see Equations (29, 30) in Tazaki & Tanaka 2018). The integration is performed with the Gauss-Legendre quadrature.

PART II.

Calculate translation matrix coefficients

Translation coefficients: A and B (Equation # from Tazaki & Tanaka 2018) T(1,nu,n) : \bar{A}{1,n}^{1,nu} defined in Equation (14) T(2,nu,n) : \bar{A}{1,n}^{1,nu} defined in Equation (15) anunp : a(nu,n,p) defined in Equation (29) bnunp : b(nu,n,p) defined in Equation (30)

Note about the rule of rum with respect to the index p (RT and Botet, in private communication, 2017).

Just before Equation (4) in Botet et al. (1997), it is written that p should have the same parity as n+nu. However, this is only true true for a(nu,n,p), and not for b(nu,n,p) Thus, in this code, index p runs ALL integers between |nu-n| and nu+n.

Tips for efficient computation. The parity of the integrand of a and b leads following properties:

a(nu,n,p) .ne. 0 and b(nu,n,p) = 0 when p has the same parity as n+nu

b(nu,n,p) .ne. 0 and a(nu,n,p) = 0 when p has the opposite parity to n+nu