kappa::fractal

Function mean_scattering

pub fn mean_scattering(fracc: &FractalConfig) -> Result<FractalResult>

Computes light scattering properties of randomly oriented fractal dust aggregates by means of the modified mean field theory developed in Tazaki & Tanaka (2018). This code is also capable of computing the light scattering solution based on the Rayleigh-Gans-Debye theory and the Mean field theory.

Arguments

iqsca: Method switch for the light scatteing solver

iqsca = 1: Rayleigh-Gans-Debye theory

iqsca = 2: Mean field theory

iqsca = 3: Modified mean-field theory

iqcor: Switch for the two-point correction function

iqcor = 1: Gaussian cut-off

iqcor = 2: Exponential cut-off

iqcor = 3: Fractal dimension cut-off

iqgeo: Switch for the geometric cross-section

iqgeo = 1: π * rc^2, where rc is the characteristic radius.

iqgeo = 2: Okuzumi et al. (2009)

iqgeo = 3: Tazaki (2021)

nang: Number of angles (0, π/2) for the scattering phase function.

pn: Number of monomers in the aggregate.

r0: Monomer radius.

df: Fractal dimension.

k: Prefactor for the fractal dimension.

lmd: Wavelength of the incident light.

refrel: Complex Refractive index of the monomer.

Notes

For iqsca = 1: all outputs would be physically reasonable for the phase shift < ~1.

For iqsca = 2: The extinction cross section would be calculated without limitation. However, the other outputs would be reliable for the phase shift < ~1.

For iqsca = 3: The extinction cross section would be calculated without limitation. Scattering and absorption cross sections could be calculated for the phase shift > 1, however too large phase shift may cause some problem. The asymmetry parameter and the sattering matrix elements would be reliable for the phase shift < ~1.

For the two-point correlation function, it is suggested to use iqcor = 1 because it is numerically stable and is likely to reproduce optical properties of fractal aggregates.

References

iqsca = 1; iqcor = 1: Tazaki et al. (2016), ApJ, 823, 70

iqsca = 2: Botet et al. (1997), ApOpt, 36, 8791

iqsca = 3: Tazaki & Tanaka (2018), ApJ, 860, 79

iqcor = 2: Berry & Percival (1986), AcOpt, 33, 577

iqcor = 3: Botet et al. (1995), JPhA, 28, 297

iqgeo = 2: Okuzumi et al. (2009), ApJ, 707, 1247

iqgeo = 3: Tazaki (2021), MNRAS, 504, 2811