kappa::fractal

Function structure_factor_integration

fn structure_factor_integration(d: f64, q: f64, r_g: f64) -> f64

This subroutine performs integration of the static structure factor by Botet et al. (1995) based on df-dependent cut-off function for two-point correlation function. The statis structure factor S(q) is given by

$$ S(q) = \frac{c \cdot df}{q R_g} \int_0^{x_{\max}} dx , x^{df-2} \sin(q R_g x) e^{-c x^{df}} $$

where x = u/Rg; where Rg is the radius of gyration and u is the distance between two monomers; q is the magnitude of scattering vector, df is the fractal dimension, c=0.5 is normalization factor.

Since we have a steep cut-off function for x>1, it would be enough to take x_max slightly larger than unity. Here we determine xmax so that eta=c*x^df becomes 25.

We use Simpson’s rule for the numerical integration, and then, the integration range [0,xmax] will be divided by 2*m points.

Arguments

d: Fractal dimension.

q: Magnitude of scattering vector.

r_g: Radius of gyration.

Returns

Value of the static structure factor.

Caution

For large df(>2) and large qRg, the numerical integration sometimes returns a negative value of S(q). Since S(q) is a power spectrum, it must be positive. I also noticed that the same problem arises with the code written by Pascal Rannou. Because of this ill behavior, I prefer to use Gausian-cut-off model, which is more stable.

A not-too-bad tentative prescription is simply set S(q)=0 when S(q)<0, which means scattered waves from each monomer is incoherent. This is what I did in if-sentence at the end of this subroutine.