kappa

Module geofractal

Enums

AFactor

IntegrationMethod

Method for solving radial and angular integration when computing the mean overlapping efficiency.

Functions

get_geometric_cross_section_tazaki

Computing geometrical cross-section of randomly oriented fractal dust aggregates based on a statistical distribution model of monomer particles as discussed in Tazaki (2021).