kappa

Module opac

The heart of kappa. This module contains the main routines to compute opacities and scattering matrices.

Structs

Component

Defines a material component which is initialized at runtime by consuming a LNK file.

KappaConfig

Material

Mueller

Mueller matrix structure

Particle

Particle

Enums

KappaError

KappaMethod

MaterialKind

RefractiveIndexKind

SizeDistribution

WavelengthKind

Traits

SpecialConfigs

Functions

run

Run the simulation.