kappa::opac

Struct Component

pub struct Component {
    pub name: String,
    pub class: String,
    pub state: String,
    pub rho: f64,
    pub size: usize,
    pub l0: RVector,
    pub n0: RVector,
    pub k0: RVector,
}

Defines a material component which is initialized at runtime by consuming a LNK file.

See crate::components::StaticComponent for an equivalent version which can is instantiated statically at compile time.

Fields
name: String

Name of the component.

class: String

Class of the component.

state: String

State of the component.

rho: f64

Density of the component.

size: usize

Number of wavelengths.

l0: RVector

Wavelengths.

n0: RVector

Refractive index.

k0: RVector

Extinction coefficient.