kappa::utils::bessel

Function int_sph_bessel

pub fn int_sph_bessel(
    fracc: &FractalConfig,
    x_g: f64,
    p: usize,
) -> Result<Complex64>

This subroutine performs integration of S_p(kRg) (Equation 31): $$ S_p(k R_g) = \frac{\pi^2}{k^3} \int_0^{\infty} du , u , J_{p+1/2}(u) , H_{p+1/2}^{(1)}(u) , g(u/k) $$
where g(u) is the two-point correlation function (Equations 18 and 19): $$ g(u) = \frac{1}{4 \pi R_g^3} \left( \frac{u}{R_g} \right)^{d_f - 3} \cdot fc \left( \frac{u}{R_g} \right) $$ where fc is the cut-off function. By substituting g(u) into S_p, we have $$ S_p(k R_g) = \frac{\pi}{4 X_g^{d_f}} \int_{u_{\min}}^{u_{\max}} du , u^{d_f - 2} J_{p+1/2}(u) H_{p+1/2}^{(1)}(u) fc(u / X_g) $$

where the integration range is approximated by the range [u_min,u_max]. By using the spherical Bessel j_p(u) and Hankel functions of 1st kind h_p^(1)(u), the Bessel function and the Hankel function are rewritten by $$ J_{p+1/2} (u) = \sqrt{\frac{2u}{\pi}} , j_p(u) $$

$$ H_{p+1/2}^{(1)}(u) = \sqrt{\frac{2u}{\pi}} , h_p^{(1)}(u) $$

We have:

$$ S_p(k R_g) = \frac{1}{2 X_g^{d_f}} \int_{u_{\min}}^{u_{\max}} du , u^{d_f - 1} j_p(u) h_p^{(1)}(u) fc(u / X_g) $$

For the unitary condition of the two-point correlation function:

$$ 1 = \int_0^{\infty} dw , 4\pi w^2 g(w) $$

If we take the integration variable as ( w = u/k ), then we obtain:

$$ 1 = \frac{1}{X_g^{d_f}} \int_{u_{\min}}^{u_{\max}} du , u^{d_f - 1} fc(u / X_g) \quad \text{…. Eq. (*)} $$

The integration range [umin,umax] is determined as follows. The integrand of Equation (*) is

    (u/xg)^{df-1}fc(u/xg) du = (u/xg) ^{d_f}fc(u/xg) dlnu

u_max is chosen so that fc(u/xg) ~ exp[-eta1]. For iqcor=1 (Gauss) u_max ~ 2 * xg * sqrt(eta1 / d_f) For iqcor=1 (Exponential) u_max ~ xg * eta1 * sqrt(2.0 /(d_f*(d_f+1)) For iqcor=1 (FLDIM) u_max ~ xg * sqrt( 2.0 * eta1 ) ** (1.0/d_f) I adopt eta1 = 25.0.

u_min is chosen so that (u_min/xg)^{d_f} ~ exp[-eta2], thus,

          umin ~ xg * exp(-eta2/d_f)

where eta2 = 40.0.

Remarks

BoundaryCondition::TazakiTanaka is used for the boundary condition and is recommended; although BoundaryCondition::Jablonski is also available.