kappa::utils

Function get_sizedis_moments

pub fn get_sizedis_moments(kpc: &KappaConfig) -> [f64; 3]

Compute the moments of the size distribution

The results are returned in ameans, an array of length 3: $$ [\langle a \rangle,\quad \langle a^2 \rangle^{1/2},\quad \langle a^3 \rangle^{1/3}] $$ If both mn and sig are nonzero and the product is positive, we use the log-normal size distribution. If not, we use the powerlaw.