kappa::utils

Module spline

Functions

spline

Given the arrays xv and yv of lengh n containing a tabulated function, i.e., yv[i] = f(xv[i]), with xv[0] < xv[1] < ... < xv[n-1], and given the first derivative values yp1 and ypn for the first derivative of the interpolating function at points xv[0] and xv[n-1], this function returns the second derivative values y2 of length n which contains the second derivatives of the interpolating function at the tabulated points xv[i].

splint

Given the arrays xa and ya which tabulate a function (with the xa[i] in order), and given the array y2a, which is the output from spline, and given a value of x, this function returns a cubic-spline interpolated value y.