kappa::utils::spline

Function spline

pub fn spline(xv: &[f64], yv: &[f64], n: usize, yp1: f64, ypn: f64) -> RVector

Given the arrays xv and yv of lengh n containing a tabulated function, i.e., yv[i] = f(xv[i]), with xv[0] < xv[1] < ... < xv[n-1], and given the first derivative values yp1 and ypn for the first derivative of the interpolating function at points xv[0] and xv[n-1], this function returns the second derivative values y2 of length n which contains the second derivatives of the interpolating function at the tabulated points xv[i].

If yp1 and ypn are equal or larger than 0.99e99, the function will set the first derivatives at the boundaries to be zero.

Reference

Numerical Recipes: The Art of Scientific Computing, 3rd ed. Press et al., 2007