Crate kappa

Modules

cli

Command line interface for the kappa.

components

This module defines structs, variables and other statically defined data structures for built-in materials in kappa.

config

dhs

Distribution of Hollow Spheres (DHS) for irregular grains and low-porosity aggregates. In order to simulate irregularities in grains (irregular shapes, or the properties of low-porosity aggregates), kappa averages the opacities of grains with an inner empty region, over a range of volume fractions of this inner region between 0 and fmax. The subroutine used to compute the opacities and scattering matrix elements for these structures is toon_ackerman_1981 (Toon & Ackerman 1981). For speed, you can set xlim 1e3 or so to set a limit for the size parameter (x = 2πa/λ) where kappa switches from DHS to crate::mie::de_rooij_1984.

fractal

Computes light scattering properties of randomly oriented fractal dust aggregates by means of the modified mean field theory developed in Tazaki & Tanaka (2018). This code is also capable of computing the light scattering solution based on the Rayleigh-Gans-Debye theory and the Mean field theory.

geofractal

io

Read the lnk files

mie

Mie scattering calculation

opac

The heart of kappa. This module contains the main routines to compute opacities and scattering matrices.

types

utils