kappa

Module dhs

Distribution of Hollow Spheres (DHS) for irregular grains and low-porosity aggregates. In order to simulate irregularities in grains (irregular shapes, or the properties of low-porosity aggregates), kappa averages the opacities of grains with an inner empty region, over a range of volume fractions of this inner region between 0 and fmax. The subroutine used to compute the opacities and scattering matrix elements for these structures is toon_ackerman_1981 (Toon & Ackerman 1981). For speed, you can set xlim 1e3 or so to set a limit for the size parameter (x = 2πa/λ) where kappa switches from DHS to crate::mie::de_rooij_1984.

Structs

DHSConfig

DHSResult

Functions

toon_ackerman_1981

This function computes electromagnetic scattering by a stratified sphere, i.e., a particle consisting of a spherical core surrounded by a spherical shell. The surrounding medium is assumed to have refractive index unity. The formulas, manipulated to avoid the ill-conditioning that plagued earlier formulations, were published in: Toon, O. and T. Ackerman, Applied Optics 20, 3657 (1981).