kappa

Module fractal

Computes light scattering properties of randomly oriented fractal dust aggregates by means of the modified mean field theory developed in Tazaki & Tanaka (2018). This code is also capable of computing the light scattering solution based on the Rayleigh-Gans-Debye theory and the Mean field theory.

References

Based on OPFRACTAL v3.0 from optool by Ryo Tazaki.

Structs

FractalConfig

FractalResult

Enums

FractalCutoff

FractalGeometry

FractalSolver

Functions

mean_scattering

Computes light scattering properties of randomly oriented fractal dust aggregates by means of the modified mean field theory developed in Tazaki & Tanaka (2018). This code is also capable of computing the light scattering solution based on the Rayleigh-Gans-Debye theory and the Mean field theory.