kappa::geofractal

Function get_geometric_cross_section_tazaki

pub fn get_geometric_cross_section_tazaki(
    method: &IntegrationMethod,
    afac: &AFactor,
    cutoff: &FractalCutoff,
    pn: f64,
    k0: f64,
    df: f64,
) -> Result<f64>

Computing geometrical cross-section of randomly oriented fractal dust aggregates based on a statistical distribution model of monomer particles as discussed in Tazaki (2021).

Arguments

iqapp - Select a method for solving radial and angular integration when computing the mean overlapping efficiency.

  --------------------------------------------------------
            |    radial  (x)        |  angular (u)        |
  ---------------------------------------------------------
  iqapp = 1 |   numerical           |  numerical          |
  iqapp = 2 |   numerical  ( D<=2 ) |  analytical         |
            |   analytical ( D> 2 ) |                     |
  --------------------------------------------------------