kappa

Module opac

The heart of kappa. This module contains the main routines to compute opacities and scattering matrices.

Structs

Component

Defines a material component which is initialized at runtime by consuming a LNK file.

KappaConfig

KappaResult 🔒

KappaState 🔒

Material

Mueller

Mueller matrix structure

Particle

Particle

Enums

KappaError

KappaMethod

MaterialKind

RefractiveIndexKind

SizeDistribution

WavelengthKind

Traits

SpecialConfigs

Functions

bruggeman_blend 🔒

compute_kappa 🔒

Calculate the opacities for a grain-size value.

maxwell_garnet_blend 🔒

over_wavelengths 🔒

Calculate the opacities for a wavelength value.

prepare_inputs 🔒

Pre-process the inputs collected from the command-line

run

Run the simulation.