kappa::utils

Module constants

Defines mathematical expressions commonly used when computing distribution values as constants

This module is directly copied from the statrs crate. The original code is licensed under the MIT license. It was purely done to avoid a dependency on the statrs crate.

The original code can be found here: https://github.com/statrs-dev/statrs/blob/master/src/constants.rs

Constants

DEFAULT_F64_ACC

Default accuracy for f64, equivalent to 0.0 * F64_PREC

LN_2_SQRT_E_OVER_PI

Constant value for ln(2 * sqrt(e / pi))

LN_PI

Constant value for ln(pi)

TWO_SQRT_E_OVER_PI

Constant value for 2 * sqrt(e / pi)